Updated data types and structure
Sorry to commit directly to master but i have updated some data types,
constructors and methods to reflect the Class Diagram.

The import/export is ready for multipal bussinesses but wont impport
more then one yet, so use .get(0) on the list to get the single one that
is imported.

I changed where the owner is stored, it makes much more sense for it to
be in the Business class.

Due to the changes i had to update some tests and other sections also.